+++ src/jvmTest/kotlin/io/github/koalaplot/core/legend/ColumnLegendTest.kt
@@ -1,0 +1,20 @@
+package io.github.koalaplot.core.legend
+
+import androidx.compose.ui.test.junit4.createComposeRule
+import org.junit.Rule
+import org.junit.Test
+
+class ColumnLegendTest {
+    @get:Rule
+    val composeTestRule = createComposeRule()
+
+    /**
+     * Test creating a ColumnLegend.
+     */
+    @Test
+    fun BasicColumnLegendTest() {
+        composeTestRule.setContent {
+            ColumnLegend(5)
+        }
+    }
+}